set purge permission

+++ src/main/java/com/kojubu/jabda/commandUtils.java
@@ -1,5 +1,6 @@
 package com.kojubu.jabda;
 
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
@@ -12,7 +13,7 @@ public class commandUtils extends ListenerAdapter {
     public void onSlashCommandInteraction(SlashCommandInteractionEvent event) {
         int messageIndex = Objects.requireNonNull(event.getOption("purge_amount")).getAsInt();
         String command = event.getName();
-        if (command.equals("삭제")) {
+        if (command.equals("삭제") && event.getMember().hasPermission(Permission.MESSAGE_MANAGE)) {
             List<Message> messages = event.getChannel().getHistory().retrievePast(messageIndex).complete();
             event.deferReply().setEphemeral(false).queue();
             event.getChannel().purgeMessages(messages);
